+++ R/taxlist2df.R
@@ -10,6 +10,10 @@
 #'     attributes have to be included in the output or not.
 #' @param include_views A logical value indicating whether taxon views
 #'     have to be included in the output or not.
+#' @param standard A character value indicating the standard used to name
+#'     columns in the output data frame. Per default `taxlist` names are used
+#'     but it can be set to `dwc` for renaming some columns according to
+#'     [Darwin Core](https://dwc.tdwg.org/terms/#taxon).
 #' @param ... Further arguments passed among methods.
 #'
 #' @rdname taxlist2df
@@ -22,7 +26,12 @@ taxlist2df <- function(x, ...) UseMethod("taxlist2df", x)
 #' @export
 taxlist2df.taxlist <- function(
     x, include_traits = FALSE,
-    include_views = FALSE, ...) {
+    include_views = FALSE,
+    standard = c("taxlist", "dwc"), ...) {
+  standard <- pmatch(standard[1], c("taxlist", "dwc"))
+  if (is.na(standard)) {
+    stop("The provided value for argument 'standard' is not valid.")
+  }
   x <- as(x, "list")
   double_names <- names(x$taxonNames)[names(x$taxonNames) %in%
     names(x$taxonRelations) & names(x$taxonNames) != "TaxonConceptID"]
@@ -54,6 +63,12 @@ taxlist2df.taxlist <- function(
       )
     }
     x_out <- merge(x_out, x$taxonViews, all = TRUE)
+  }
+  if (standard == 2) {
+    names(x_out) <- replace_x(names(x_out),
+      old = dwc_table$taxlist,
+      new = dwc_table$dwc
+    )
   }
   return(x_out)
 }