+++ data-raw/create-data.R
@@ -10,3 +10,9 @@ veg_layers <- read.csv("data-raw/veg_layers.csv")
 veg_layers$layer <- as.factor(veg_layers$layer)
 
 save(veg_layers, file = "data/veg_layers.rda")
+
+# Darwin Core names ------------------------------------------------------------
+dwc_table <- read.csv("data-raw/dwc-table.csv")
+
+# Save System data
+save(dwc_table, file = "R/sysdata.rda")